refactor(T11404): reclaim core/src/tools — relocate sdk dir, completing the reclaim (E3 · SG-PACKAGE-ARCH)
Final reclaim wave. Moves the last squatter dir out of core/src/tools:

- core/src/tools/sdk/ → core/src/sdk/

Decremented the dir's 23 escaping relative imports (../../ → ../); repointed the
4 external importers (orchestrate/spawn-ops, orchestration/spawn-prompt,
validation/validate-ops, lifecycle validator integration test) from
../tools/sdk → ../sdk; fixed the mirror test under tools/__tests__/sdk/; updated
the stale contracts/sdk-tool.ts doc comment; repointed the tools/index.ts barrel.

core/src/tools now holds ONLY the atomic primitives (fs/shell/guard) + the
barrel — ALL 8 squatters (5 files + 3 dirs) reclaimed to honest core siblings.
@cleocode/core public API is UNCHANGED (zero behavior change).

Verified: core build (tsc) · cleo build (public API intact) · cleo check arch
5/5 · deprecations + tools-vs-skills boundary clean · mirror-test imports
resolve · biome.
