+++ packages/core/src/orchestration/spawn-prompt.ts
@@ -43,8 +43,8 @@ import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { Task } from '@cleocode/contracts';
 import { normalizeSlug } from '../docs/slug-normalize.js';
+import { ISOLATION_ENV_KEYS, provisionIsolatedShell } from '../sdk/isolation.js';
 import { resolveSkillPath } from '../skills/skill-paths.js';
-import { ISOLATION_ENV_KEYS, provisionIsolatedShell } from '../tools/sdk/isolation.js';
 
 /**
  * Locate `packages/core/templates/CLEO-INJECTION.md` at runtime.
