+++ packages/core/src/sdk/validator-attest.ts
@@ -25,9 +25,9 @@ import {
   type ValidatorAttestation,
   validatorAttestationSchema,
 } from '@cleocode/contracts';
-import { getTaskAccessor } from '../../store/data-accessor.js';
-import type { JsonSchema, RegisteredSdkTool } from '../../task-tools/sdk-tool.js';
-import { defineSdkTool } from '../../task-tools/sdk-tool.js';
+import { getTaskAccessor } from '../store/data-accessor.js';
+import type { JsonSchema, RegisteredSdkTool } from '../task-tools/sdk-tool.js';
+import { defineSdkTool } from '../task-tools/sdk-tool.js';
 
 /**
  * Input envelope for the {@link validatorAttest} SDK tool.
