+++ packages/core/src/sdk/validator-reject.ts
@@ -24,8 +24,8 @@ import {
   type ValidatorRejection,
   validatorRejectionSchema,
 } from '@cleocode/contracts';
-import type { JsonSchema, RegisteredSdkTool } from '../../task-tools/sdk-tool.js';
-import { defineSdkTool } from '../../task-tools/sdk-tool.js';
+import type { JsonSchema, RegisteredSdkTool } from '../task-tools/sdk-tool.js';
+import { defineSdkTool } from '../task-tools/sdk-tool.js';
 
 /**
  * Input envelope for the {@link validatorReject} SDK tool.
